requirements.txt
kafka-python
requests
shapely
